+++ core-ng-search/src/main/java/core/framework/search/module/SearchConfig.java
@@ -7,7 +7,6 @@ import core.framework.search.ElasticSearch;
 import core.framework.search.ElasticSearchType;
 import core.framework.search.impl.ElasticSearchHost;
 import core.framework.search.impl.ElasticSearchImpl;
-import core.framework.util.Strings;
 import core.framework.util.Types;
 
 import java.time.Duration;
@@ -17,6 +16,7 @@ import java.time.Duration;
  */
 public class SearchConfig extends Config {
     ElasticSearchImpl search;
+    boolean auth;
     private ModuleContext context;
     private String name;
     private boolean typeAdded;
@@ -38,27 +38,19 @@ public class SearchConfig extends Config {
         if (search.hosts == null) throw new Error("search host must be configured, name=" + name);
         if (!typeAdded)
             throw new Error("search is configured but no type added, please remove unnecessary config, name=" + name);
-    }
-
-    // comma separated hosts
-    public void host(String host, String apiKey, String keySecret) {
-        if (!Strings.isBlank(apiKey)) search.apiKey = apiKey;
-        if (!Strings.isBlank(keySecret)) search.keySecret = keySecret;
-
-        search.hosts = ElasticSearchHost.parse(host);
-
-        // skip probe if auth has been configured, most likely it isn't deployed in kube env
-        if (!hasAuth()) {
+        if (!auth) { // skip probe if auth has been configured, most likely it isn't deployed in kube env
             context.probe.urls.add(search.hosts[0].toURI() + "/_cluster/health?local=true");      // in kube env, it's ok to just check first pod of stateful set
         }
     }
 
+    // comma separated uris
     public void host(String host) {
-        host(host, null, null);
+        search.hosts = ElasticSearchHost.parse(host);
     }
 
-    boolean hasAuth() {
-        return !Strings.isBlank(search.apiKey) && !Strings.isBlank(search.keySecret);
+    public void auth(String apiKeyId, String apiKeySecret) {
+        search.auth(apiKeyId, apiKeySecret);
+        auth = true;
     }
 
     // refer to https://www.elastic.co/guide/en/elasticsearch/reference/current/index-modules.html#index-max-result-window