+++ core-ng-search/src/test/java/core/framework/search/impl/ElasticSearchHostTest.java
@@ -4,6 +4,7 @@ import org.apache.http.HttpHost;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 /**
  * @author neo
@@ -11,22 +12,29 @@ import static org.assertj.core.api.Assertions.assertThat;
 class ElasticSearchHostTest {
     @Test
     void parse() {
-        assertThat(ElasticSearchHost.parse("es-0"))
+        assertThat(ElasticSearchHost.parse("http://es-0"))
             .containsExactly(new HttpHost("es-0", 9200));
 
-        assertThat(ElasticSearchHost.parse("es-0, es-1"))
+        assertThat(ElasticSearchHost.parse("http://es-0, http://es-1"))
             .containsExactly(new HttpHost("es-0", 9200), new HttpHost("es-1", 9200));
     }
 
     @Test
-    void parseHttps() {
-        assertThat(ElasticSearchHost.parse("https://some-secure-es.io"))
-            .containsExactly(new HttpHost("some-secure-es.io", 9200, "https"));
+    void parseWithHTTPS() {
+        assertThat(ElasticSearchHost.parse("https://es-cloud.io"))
+            .containsExactly(new HttpHost("es-cloud.io", 9200, "https"));
     }
 
     @Test
-    void customPort() {
-        assertThat(ElasticSearchHost.parse("https://some-secure-es.io:9545"))
-            .containsExactly(new HttpHost("some-secure-es.io", 9545, "https"));
+    void parseWithCustomPort() {
+        assertThat(ElasticSearchHost.parse("https://es-cloud.io:9545"))
+            .containsExactly(new HttpHost("es-cloud.io", 9545, "https"));
+    }
+
+    @Test
+    void parseWithInvalidURI() {
+        assertThatThrownBy(() -> ElasticSearchHost.parse("es-0"))
+            .isInstanceOf(Error.class)
+            .hasMessageContaining("invalid elasticsearch uri");
     }
 }